requirements.txt
flask
flask-cors
google-generativeai
gspread
oauth2client
